+++ Azure-NVMe-Utils/Changelog.md
@@ -2,6 +2,10 @@
 
 This file contains the change log for the PowerShell script.
 
+## Version 2025062704
+
+* fixing an issue where the script would not stop on Linux if there is an error during OS check
+
 ## Version 2025062703
 
 * when using Linux and not using FixOperatingSystem the script would not stop if the OS is not ready